chore: some view
clean up code and change view path
