+++ resources/js/Pages/About/Create.jsx
@@ -4,8 +4,8 @@ import { Head } from "@inertiajs/inertia-react";
 // Import Layouts
 import AuthenticatedLayout from "@/Layouts/AuthenticatedLayout";
 
-// Import Custom Components
-import CreateForm from "@/Components/About/CreateForm";
+// Import Partials
+import CreateForm from "@/Partials/About/CreateForm";
 
 export default function AboutCreate(props) {
     return (
