+++ resources/js/Pages/Portofolio/Index.jsx
@@ -4,13 +4,13 @@ import { Head } from "@inertiajs/inertia-react";
 // Import Layouts
 import PortofolioLayout from "@/Layouts/PortofolioLayout";
 
-// Import Custom Components
-import About from "@/Components/Portofolio/About";
-import Contact from "@/Components/Portofolio/Contact";
-import Projects from "@/Components/Portofolio/Projects";
-import Jumbotron from "@/Components/Portofolio/Jumbotron";
+// Import Partials
+import About from "@/Partials/Portofolio/About";
+import Contact from "@/Partials/Portofolio/Contact";
+import Projects from "@/Partials/Portofolio/Projects";
+import Jumbotron from "@/Partials/Portofolio/Jumbotron";
 
-export default function Portofolio(props) {
+export default function PortofolioIndex(props) {
     return (
         <PortofolioLayout owner={props.about} application={props.application}>
             <Head title={props.about.name} />
