+++ resources/js/Pages/Project/Index.jsx
@@ -6,7 +6,9 @@ import AuthenticatedLayout from "@/Layouts/AuthenticatedLayout";
 
 // Import Custom Components
 import Paginator from "@/Components/Paginator";
-import DataTable from "@/Components/Project/DataTable";
+
+// Import Partials
+import DataTable from "@/Partials/Project/DataTable";
 
 export default function ProjectIndex(props) {
     return (
