+++ resources/js/Partials/Translate/DataTable.jsx
