+++ resources/js/Partials/Translate/EditForm.jsx
@@ -8,15 +8,15 @@ import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 
 // Import Custom Components
-import InputError from "../InputError";
+import InputError from "../../Components/InputError";
 
-export default function CreateForm({ errors }) {
+export default function EditForm({ translate, errors }) {
     const [validated, setValidated] = useState(false);
     const [values, setValues] = useState({
-        group: "",
-        key: "",
-        id: "",
-        en: "",
+        group: translate.group,
+        key: translate.key,
+        id: translate.text.id,
+        en: translate.text.en,
     });
 
     const handleChanges = (e) => {
@@ -35,7 +35,13 @@ export default function CreateForm({ errors }) {
         }
 
         setValidated(true);
-        Inertia.post("/translate", values);
+
+        updateTransData(
+            translate.group + "." + translate.key,
+            values.group + "." + values.key
+        );
+
+        Inertia.patch("/translate/" + translate.id, values);
     };
 
     return (
@@ -49,6 +55,7 @@ export default function CreateForm({ errors }) {
                     name="group"
                     placeholder={transData("form.translate.placeholder.group")}
                     onChange={handleChanges}
+                    defaultValue={translate.group}
                     required
                     autoFocus
                 />
@@ -63,6 +70,7 @@ export default function CreateForm({ errors }) {
                     name="key"
                     placeholder={transData("form.translate.placeholder.key")}
                     onChange={handleChanges}
+                    defaultValue={translate.key}
                     required
                     autoFocus
                 />
@@ -77,6 +85,7 @@ export default function CreateForm({ errors }) {
                     name="id"
                     placeholder={transData("form.translate.placeholder.id")}
                     onChange={handleChanges}
+                    defaultValue={translate.text.id}
                     required
                     autoFocus
                 />
@@ -91,6 +100,7 @@ export default function CreateForm({ errors }) {
                     name="en"
                     placeholder={transData("form.translate.placeholder.en")}
                     onChange={handleChanges}
+                    defaultValue={translate.text.en}
                     required
                     autoFocus
                 />
